# sf.Server/Properties/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Register",
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Login",
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/class/{id}/school": {
      "get": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_GetSchool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfSchool"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/class/school/{schoolId}": {
      "get": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_GetBySchool",
        "parameters": [
          {
            "name": "schoolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            },
            "x-position": 2
          },
          {
            "name": "entities",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            },
            "x-position": 3
          },
          {
            "name": "properties",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "sendNull",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 5
          },
          {
            "name": "filters",
            "in": "query",
            "schema": {
              "type": "object",
              "nullable": true,
              "additionalProperties": {
                "type": "string"
              }
            },
            "x-position": 6
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfPageOfClass"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/class/{id}/tutor": {
      "get": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_GetTutor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfTutor"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/class/tutor/{tutorId}": {
      "get": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_GetByTutor",
        "parameters": [
          {
            "name": "tutorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfClass"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/class/{id}/students": {
      "get": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_GetStudents",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "x-position": 2
          },
          {
            "name": "entities",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "x-position": 3
          },
          {
            "name": "properties",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "sendNull",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "x-position": 5
          }
        ],
        "requestBody": {
          "x-name": "filters",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          },
          "x-position": 6
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfPageOfStudent"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/class": {
      "get": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_BulkGet",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            },
            "x-position": 1
          },
          {
            "name": "entities",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            },
            "x-position": 2
          },
          {
            "name": "properties",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "sendNull",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 4
          },
          {
            "name": "filters",
            "in": "query",
            "schema": {
              "type": "object",
              "nullable": true,
              "additionalProperties": {
                "type": "string"
              }
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfPageOfClass"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_Post",
        "parameters": [
          {
            "name": "ignoreNullProperties",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "entity",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Class"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfClass"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_Put",
        "parameters": [
          {
            "name": "ignoreNullProperties",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "entity",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Class"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfClass"
                }
              }
            }
          }
        }
      },
      "options": {
        "tags": [
          "Class"
        ],
        "operationId": "Class_Options",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "includeModels",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 2
          },
          {
            "name": "includeHttpCodes",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultModelOfObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "schoolname": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "LoginModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "schoolID": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
      "User": {
        "allOf": [
          {
            "$ref": "#/components/schemas/IdentityUserOfGuid"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "created",
              "updated"
            ],
            "properties": {
              "first": {
                "type": "string",
                "maxLength": 256,
                "nullable": true
              },
              "last": {
                "type": "string",
                "maxLength": 256,
                "nullable": true
              },
              "role": {
                "$ref": "#/components/schemas/RoleType"
              },
              "created": {
                "type": "string",
                "format": "date-time",
                "minLength": 1
              },
              "updated": {
                "type": "string",
                "format": "date-time",
                "minLength": 1
              }
            }
          }
        ]
      },
      "RoleType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Student",
          "Tutor",
          "CampaignManager",
          "CampaignJudge",
          "User"
        ],
        "enum": [
          "Student",
          "Tutor",
          "CampaignManager",
          "CampaignJudge",
          "User"
        ]
      },
      "IdentityUserOfGuid": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ResultModelOfSchool": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/School"
              }
            ]
          },
          "messages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "School": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EntityOfGuid"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 64,
                "minLength": 2
              },
              "shortName": {
                "type": "string",
                "maxLength": 16,
                "minLength": 2
              },
              "comment": {
                "type": "string",
                "maxLength": 1024
              },
              "managerId": {
                "type": "string",
                "format": "guid",
                "nullable": true
              },
              "judgeIds": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/Tutor"
                }
              },
              "studentIds": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        ]
      },
      "Tutor": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "classId": {
                "type": "string",
                "format": "guid",
                "nullable": true
              },
              "schoolId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "EntityOfGuid": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "created",
          "updated"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          }
        }
      },
      "Message": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error.message": {
            "type": "string",
            "nullable": true
          },
          "error.type": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ErrorResult"
              }
            ]
          },
          "error.severity": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Severity"
              }
            ]
          }
        }
      },
      "ErrorResult": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "InternalException",
          "InvalidModel",
          "InvalidCredentials",
          "InvalidToken",
          "InvalidRole",
          "InvalidUser",
          "InvalidPassword",
          "InvalidEmail",
          "InvalidUsername",
          "InvalidOldPassword",
          "InvalidNewPassword",
          "InvalidModelState",
          "PasswordChangeFailed",
          "AccountLocked",
          "UserNotFound",
          "UserCreationFailed",
          "RoleCreationFailed",
          "Validation",
          "InvalidId"
        ],
        "enum": [
          "InternalException",
          "InvalidModel",
          "InvalidCredentials",
          "InvalidToken",
          "InvalidRole",
          "InvalidUser",
          "InvalidPassword",
          "InvalidEmail",
          "InvalidUsername",
          "InvalidOldPassword",
          "InvalidNewPassword",
          "InvalidModelState",
          "PasswordChangeFailed",
          "AccountLocked",
          "UserNotFound",
          "UserCreationFailed",
          "RoleCreationFailed",
          "Validation",
          "InvalidId"
        ]
      },
      "Severity": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Verbose",
          "Debug",
          "Information",
          "Warning",
          "Error",
          "Fatal"
        ],
        "enum": [
          "Verbose",
          "Debug",
          "Information",
          "Warning",
          "Error",
          "Fatal"
        ]
      },
      "ResultModelOfPageOfClass": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PageOfClass"
              }
            ]
          },
          "messages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "PageOfClass": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          }
        }
      },
      "Class": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EntityOfGuid"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 64
              },
              "short": {
                "type": "string",
                "maxLength": 8
              },
              "comment": {
                "type": "string",
                "maxLength": 1024
              },
              "tutorId": {
                "type": "string",
                "format": "guid",
                "nullable": true
              },
              "roomId": {
                "type": "string",
                "format": "guid",
                "nullable": true
              },
              "schoolId": {
                "type": "string",
                "format": "guid",
                "nullable": true
              },
              "locationId": {
                "type": "string",
                "format": "guid",
                "nullable": true
              }
            }
          }
        ]
      },
      "ResultModelOfTutor": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Tutor"
              }
            ]
          },
          "messages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "ResultModelOfClass": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Class"
              }
            ]
          },
          "messages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "ResultModelOfPageOfStudent": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PageOfStudent"
              }
            ]
          },
          "messages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "PageOfStudent": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Student"
            }
          }
        }
      },
      "Student": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "classId": {
                "type": "string",
                "format": "guid",
                "nullable": true
              },
              "schoolId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
...
      "ResultModelOfObject": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          },
          "messages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
...
    }
  }
}